+++ src/Regions.jsx
@@ -1,4 +1,4 @@
-export default function Regions({ regions }) {
+export default function Regions({ regions, onClick }) {
   if (regions.length === 0) {
     return (
       <p>지역이 없습니다.</p>
@@ -9,7 +9,13 @@ export default function Regions({ regions }) {
     <ul>
       {regions.map((region) => (
         <li key={region.id}>
-          <button type="button">{region.name}</button>
+          <button
+            type="button"
+            id={region.id}
+            onClick={onClick}
+          >
+            {region.name}
+          </button>
         </li>
       ))}
     </ul>
